refactor: too many things, refactored the form and functions and added better step and error handling
