+++ public/installer/views/admin-creation.php
@@ -6,10 +6,10 @@
     $subtitle = "Lets create the first admin user!"
 ); ?>
 
-<form method="POST" enctype="multipart/form-data" class="m-0" action="/installer/forms.php" name="createUser">
+<form method="POST" enctype="multipart/form-data" class="m-0" action="/installer/index.php" name="createUser">
 
-    <?php if (isset($_GET['message'])) {
-        echo "<p class='not-ok check'>" . $_GET['message'] . '</p>';
+    <?php if (isset($_SESSION['error-message'])) {
+        echo "<p class='not-ok check'>" . $_SESSION['error-message'] . '</p>';
     } ?>
 
     <div class="form-group">
