+++ public/installer/views/dashboard-configuration.php
@@ -5,10 +5,10 @@
     $title = "Dashboard Configuration"
 ); ?>
 
-<form method="POST" enctype="multipart/form-data" class="m-0" action="/installer/forms.php" name="checkGeneral">
+<form method="POST" enctype="multipart/form-data" class="m-0" action="/installer/index.php" name="checkGeneral">
 
-    <?php if (isset($_GET['message'])) {
-        echo "<p class='not-ok check'>" . $_GET['message'] . '</p>';
+    <?php if (isset($_SESSION['error-message'])) {
+        echo "<p class='not-ok check'>" . $_SESSION['error-message'] . '</p>';
     } ?>
 
     <div class="row">
