+++ public/installer/views/pterodactyl-configuration.php
@@ -6,10 +6,10 @@
     $subtitle = "Lets get some info about your Pterodactyl Installation!"
 ); ?>
 
-<form method="POST" enctype="multipart/form-data" class="m-0" action="/installer/forms.php" name="checkPtero">
+<form method="POST" enctype="multipart/form-data" class="m-0" action="/installer/index.php" name="checkPtero">
 
-    <?php if (isset($_GET['message'])) {
-        echo "<p class='not-ok check'>" . $_GET['message'] . '</p>';
+    <?php if (isset($_SESSION['error-message'])) {
+        echo "<p class='not-ok check'>" . $_SESSION['error-message'] . '</p>';
     } ?>
 
     <div class="row">
